requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.23.0
google-cloud-spanner==3.54.0
google-auth==2.10.0
pydantic==1.10.0
gunicorn
Jinja2==3.1.2
